finalize result component
